+++ src/components/CreationCardMinimal/CreaitonCardMinimal.tsx
