+++ src/components/EdenArtFrontPage/EdenArtFrontPage.tsx
