+++ src/components/VideoResult/VideoResult.tsx
